refactor: extract DeepSeek endpoint to standalone deepseek_.py
Move DeepseekChatEndpoint and _get_deepseek_config from oai_.py to
providers/deepseek_.py. Matches the pattern of other non-OAI providers
(anthropic_, perplexity_, etc.). Update match_endpoint and test imports.
