+++ lionagi/service/connections/providers/oai_.py
@@ -19,10 +19,6 @@ from pydantic import BaseModel
 from lionagi.config import settings
 from lionagi.service.connections.endpoint import Endpoint
 from lionagi.service.connections.endpoint_config import EndpointConfig
-from lionagi.service.third_party.deepseek_models import (
-    DeepseekChatCompletionsRequest,
-    normalize_deepseek_usage,
-)
 from lionagi.service.third_party.openai_models import OpenAIChatCompletionsRequest
 
 
@@ -146,52 +142,6 @@ class GeminiChatEndpoint(Endpoint):
     def __init__(self, config=None, **kwargs):
         config = config or _get_gemini_config()
         super().__init__(config, **kwargs)
-
-
-def _get_deepseek_config(**kwargs):
-    """Create DeepSeek endpoint configuration with defaults."""
-    config = dict(
-        name="deepseek_chat",
-        provider="deepseek",
-        base_url="https://api.deepseek.com/v1",
-        endpoint="chat/completions",
-        kwargs={"model": "deepseek-chat"},
-        api_key=settings.DEEPSEEK_API_KEY or "dummy-key-for-testing",
-        auth_type="bearer",
-        content_type="application/json",
-        method="POST",
-        request_options=DeepseekChatCompletionsRequest,
-    )
-    config.update(kwargs)
-    return EndpointConfig(**config)
-
-
-class DeepseekChatEndpoint(Endpoint):
-    def __init__(self, config=None, **kwargs):
-        config = config or _get_deepseek_config()
-        super().__init__(config, **kwargs)
-
-    def create_payload(
-        self,
-        request: dict | BaseModel,
-        extra_headers: dict | None = None,
-        **kwargs,
-    ):
-        payload, headers = super().create_payload(request, extra_headers, **kwargs)
-        # Validate top-level DeepSeek fields (thinking, reasoning_effort)
-        # but preserve original messages — re-serializing through the model
-        # would strip DeepSeek-specific fields like reasoning_content from
-        # assistant messages that the OpenAI base model doesn't define.
-        original_messages = payload.get("messages")
-        req = DeepseekChatCompletionsRequest.model_validate(payload)
-        payload = req.model_dump(exclude_none=True, mode="json")
-        if original_messages is not None:
-            payload["messages"] = original_messages
-        return payload, headers
-
-    async def _call(self, payload: dict, headers: dict, **kwargs):
-        response = await super()._call(payload, headers, **kwargs)
-        return normalize_deepseek_usage(response)
 
 
 class OpenaiEmbedEndpoint(Endpoint):
